Added size_t dt to function

+++ 0x13-more_singly_linked_lists/0-print_listint.c
@@ -8,7 +8,7 @@
  */
 size_t print_listint(const listint_t *h)
 {
-	unsigned int node_count;
+	size_t node_count;
 	const listint_t *copy;
 
 	node_count = 0;
